Lexeme: ulong -> int

+++ include/qubiq/lexeme.h
@@ -16,9 +16,9 @@ private:
     /* Each lexeme is represented in a text as a set of its forms
      * occuring in certain text positions, counted as offsets relative to
      * the first word in the text: */
-    QVector<QString>  *_forms;
-    QVector<ulong>    *_offsets;
-    QHash<ulong, int> *_idx_offsets; // Ensure that offsets are unique
+    QVector<QString> *_forms;
+    QVector<int>     *_offsets;
+    QHash<int, int>  *_idx_offsets; // Ensure that offsets are unique
 
 public:
     Lexeme(const QString &lexeme, bool is_boundary);
@@ -28,10 +28,10 @@ public:
     inline bool    isBoundary() const { return _is_boundary; }
     inline bool    isVirtual()  const { return _forms->length() == 0; }
 
-    inline const QVector<QString>* forms()   const { return _forms; }
-    inline const QVector<ulong>*   offsets() const { return _offsets; }
+    inline const QVector<QString>* forms() const { return _forms; }
+    inline const QVector<int>*   offsets() const { return _offsets; }
 
-    bool addForm(const QString &form, ulong offset, bool overwrite = false);
+    bool addForm(const QString &form, int offset, bool overwrite = false);
 };
 
 #endif // _LEXEME_H_
